+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Updated
 - Update API errors code
+- Remove the auto_create_fields parameter from index method
 
 ## [1.0.0]
 ### Added